# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Gmail Checker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface EmailCheckInput {
  /**
     * List of email addresses to verify
     * @maxItems 500
     */
  emails: string[];
}

/**
 * valid = mailbox exists; invalid = mailbox doesn't exist; disabled = account exists but is disabled/inactive; catch_all = domain accepts all; unknown = couldn't determine
 */
export type EmailResultStatus = typeof EmailResultStatus[keyof typeof EmailResultStatus];


export const EmailResultStatus = {
  valid: 'valid',
  invalid: 'invalid',
  disabled: 'disabled',
  catch_all: 'catch_all',
  unknown: 'unknown',
} as const;

export interface EmailResult {
  email: string;
  /** valid = mailbox exists; invalid = mailbox doesn't exist; disabled = account exists but is disabled/inactive; catch_all = domain accepts all; unknown = couldn't determine */
  status: EmailResultStatus;
  /** Human-readable explanation */
  reason: string;
  isGmail?: boolean;
  /**
     * SMTP response code from RCPT TO check
     * @nullable
     */
  smtpCode?: number | null;
}

export interface EmailCheckResponse {
  results: EmailResult[];
  total: number;
  valid: number;
  invalid: number;
  disabled: number;
  catchAll: number;
  unknown: number;
}

export interface EmailStatsInput {
  results: EmailResult[];
}

export interface EmailStats {
  total: number;
  valid: number;
  invalid: number;
  disabled: number;
  catchAll: number;
  unknown: number;
  validPercent: number;
  invalidPercent: number;
}

export interface LoginCredential {
  email: string;
  password: string;
  /** Base32 TOTP secret from the authenticator app setup (optional — for 2FA accounts) */
  totp?: string;
}

export interface LoginCheckInput {
  /** @maxItems 200 */
  credentials: LoginCredential[];
}

/**
 * accessible = login OK; verification_required = browser verification needed; wrong_password = bad credentials; app_password_required = 2FA on, use App Password; unknown = couldn't determine
 */
export type LoginResultStatus = typeof LoginResultStatus[keyof typeof LoginResultStatus];


export const LoginResultStatus = {
  accessible: 'accessible',
  verification_required: 'verification_required',
  wrong_password: 'wrong_password',
  app_password_required: 'app_password_required',
  unknown: 'unknown',
} as const;

export interface LoginResult {
  email: string;
  /** accessible = login OK; verification_required = browser verification needed; wrong_password = bad credentials; app_password_required = 2FA on, use App Password; unknown = couldn't determine */
  status: LoginResultStatus;
  reason: string;
  /**
     * Auto-generated TOTP code (if secret was provided)
     * @nullable
     */
  totpCode?: string | null;
  /**
     * Seconds until the TOTP code expires
     * @nullable
     */
  totpSecondsLeft?: number | null;
}

export interface LoginCheckResponse {
  results: LoginResult[];
  total: number;
  accessible: number;
  verificationRequired: number;
  wrongPassword: number;
  appPasswordRequired: number;
  unknown: number;
}

export interface ErrorResponse {
  error: string;
}