# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gmail Checker API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: emails
    description: Email verification operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /emails/check:
    post:
      operationId: checkEmails
      tags: [emails]
      summary: Bulk email verification
      description: Verifies a list of email addresses via format, MX, and SMTP checks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailCheckInput"
      responses:
        "200":
          description: Check results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailCheckResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /emails/login-check:
    post:
      operationId: loginCheckEmails
      tags: [emails]
      summary: Gmail login check via IMAP
      description: Attempts to log in to Gmail accounts via IMAP and returns accessible/verification_required/wrong_password status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginCheckInput"
      responses:
        "200":
          description: Login check results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginCheckResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /emails/stats:
    post:
      operationId: getEmailStats
      tags: [emails]
      summary: Get stats for a set of email check results
      description: Returns aggregate statistics from email check results
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailStatsInput"
      responses:
        "200":
          description: Statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    EmailCheckInput:
      type: object
      required:
        - emails
      properties:
        emails:
          type: array
          items:
            type: string
          maxItems: 500
          description: List of email addresses to verify

    EmailResult:
      type: object
      required:
        - email
        - status
        - reason
      properties:
        email:
          type: string
        status:
          type: string
          enum: [valid, invalid, disabled, catch_all, unknown]
          description: "valid = mailbox exists; invalid = mailbox doesn't exist; disabled = account exists but is disabled/inactive; catch_all = domain accepts all; unknown = couldn't determine"
        reason:
          type: string
          description: Human-readable explanation
        isGmail:
          type: boolean
        smtpCode:
          type: ["integer", "null"]
          description: SMTP response code from RCPT TO check

    EmailCheckResponse:
      type: object
      required:
        - results
        - total
        - valid
        - invalid
        - disabled
        - catchAll
        - unknown
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/EmailResult"
        total:
          type: integer
        valid:
          type: integer
        invalid:
          type: integer
        disabled:
          type: integer
        catchAll:
          type: integer
        unknown:
          type: integer

    EmailStatsInput:
      type: object
      required:
        - results
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/EmailResult"

    EmailStats:
      type: object
      required:
        - total
        - valid
        - invalid
        - disabled
        - catchAll
        - unknown
        - validPercent
        - invalidPercent
      properties:
        total:
          type: integer
        valid:
          type: integer
        invalid:
          type: integer
        disabled:
          type: integer
        catchAll:
          type: integer
        unknown:
          type: integer
        validPercent:
          type: number
        invalidPercent:
          type: number

    LoginCheckInput:
      type: object
      required:
        - credentials
      properties:
        credentials:
          type: array
          maxItems: 200
          items:
            $ref: "#/components/schemas/LoginCredential"

    LoginCredential:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
        password:
          type: string
        totp:
          type: string
          description: Base32 TOTP secret from the authenticator app setup (optional — for 2FA accounts)

    LoginResult:
      type: object
      required:
        - email
        - status
        - reason
      properties:
        email:
          type: string
        status:
          type: string
          enum: [accessible, verification_required, wrong_password, app_password_required, unknown]
          description: "accessible = login OK; verification_required = browser verification needed; wrong_password = bad credentials; app_password_required = 2FA on, use App Password; unknown = couldn't determine"
        reason:
          type: string
        totpCode:
          type: ["string", "null"]
          description: Auto-generated TOTP code (if secret was provided)
        totpSecondsLeft:
          type: ["integer", "null"]
          description: Seconds until the TOTP code expires

    LoginCheckResponse:
      type: object
      required:
        - results
        - total
        - accessible
        - verificationRequired
        - wrongPassword
        - appPasswordRequired
        - unknown
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/LoginResult"
        total:
          type: integer
        accessible:
          type: integer
        verificationRequired:
          type: integer
        wrongPassword:
          type: integer
        appPasswordRequired:
          type: integer
        unknown:
          type: integer

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string