# lib/api-zod/src/generated/types/loginCheckInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Gmail Checker API
 * OpenAPI spec version: 0.1.0
 */
import type { LoginCredential } from './loginCredential';

export interface LoginCheckInput {
  /** @maxItems 200 */
  credentials: LoginCredential[];
}