# lib/api-zod/src/generated/types/loginCheckResponse.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Gmail Checker API
 * OpenAPI spec version: 0.1.0
 */
import type { LoginResult } from './loginResult';

export interface LoginCheckResponse {
  results: LoginResult[];
  total: number;
  accessible: number;
  verificationRequired: number;
  wrongPassword: number;
  appPasswordRequired: number;
  unknown: number;
}