# lib/api-zod/src/generated/types/loginCredential.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Gmail Checker API
 * OpenAPI spec version: 0.1.0
 */

export interface LoginCredential {
  email: string;
  password: string;
  /** Base32 TOTP secret from the authenticator app setup (optional — for 2FA accounts) */
  totp?: string;
}