# lib/api-zod/src/generated/types/loginResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Gmail Checker API
 * OpenAPI spec version: 0.1.0
 */
import type { LoginResultStatus } from './loginResultStatus';

export interface LoginResult {
  email: string;
  /** accessible = login OK; verification_required = browser verification needed; wrong_password = bad credentials; app_password_required = 2FA on, use App Password; unknown = couldn't determine */
  status: LoginResultStatus;
  reason: string;
  /**
     * Auto-generated TOTP code (if secret was provided)
     * @nullable
     */
  totpCode?: string | null;
  /**
     * Seconds until the TOTP code expires
     * @nullable
     */
  totpSecondsLeft?: number | null;
}